Manifest

From meta-iotqa rev: 4cc4f1db4bb5a3933d0024051caca98f842d4eae

+++ meta-iotqa/classes/test-iot.bbclass
@@ -131,15 +131,22 @@ def dump_builddata(d, tdir):
     with open(bdpath, "w+") as f:
         json.dump(savedata, f, skipkeys=True, indent=4, sort_keys=True)
 
-#copy test description file
-def copy_testdesc(d, tdir):
-    import shutil
-    
-    descfile = "testdesc.json"
+# copy src_file(relative of layerdir) to tdir
+def copy_qa_layer_file_to(d, src_file, tdir):
+    import shutil, os
     layerdir = get_qa_layer(d)
-    srcpath = os.path.join(layerdir, "conf", "test", descfile)
-    shutil.copy2(srcpath, tdir)
-    bb.plain("copy %s to: %s" % (descfile, tdir))
+    srcpath = os.path.join(layerdir, src_file)
+    if os.path.isfile(srcpath):
+        shutil.copy2(srcpath, tdir)
+        bb.plain("copy %s to: %s" % (srcpath, tdir))
+    else:
+        bb.plain("missing file: %s" % srcpath)
+
+# copy manifest file
+def copy_manifest(d, tdir):
+    srcpath = os.path.join(get_qa_layer(d), "conf", "test")
+    import shutil
+    shutil.copytree(srcpath, tdir)
 
 def re_creat_dir(path):
     bb.utils.remove(path, recurse=True)
@@ -199,7 +206,8 @@ python do_test_iot_export() {
     bb.utils.mkdirhier(os.path.join(exportdir, deploydir))
     export_testsuite(d, exportdir)
     dump_builddata(d, exportdir)
-    copy_testdesc(d, exportdir)
+    plandir = os.path.join(exportdir, "testplan")
+    copy_manifest(d, plandir)
     fname = "/tmp/iot-testsuite.tar.gz"
     pack_tarball(d, exportdir, fname)
